// src/app/customer/models/customer-real-owner.ts
export interface CustomerRealOwnerPPE {
  idM?: number;
  id?: number;
  ppe: boolean;
  tppe: string;
  positionHeld: string;
  expirationDate: string;
  fppe: boolean;
  dataFamily: Array<DataRealOwnerClientFamilyPPESave>;
}

export interface DataRealOwnerClientFamilyPPE
  extends CustomerMinRealOwnerDataClient,
  CustomerPositionHeld {
  idM?: number;
  id?: string | number;
  idS?: string;
  accountRoleId?: number;
  active?: boolean;
  personId?: number;
  isView?: boolean;
  isSaved?: boolean;
}

interface DataRealOwnerClientFamilyPPESave
  extends CustomerMinRealOwnerDataClient,
  CustomerPositionHeld {
  idM?: number;
  id?: string | number;
  idS?: string;
  personId?: number;
  active?: boolean;
  accountRoleId?: number;
  isView?: boolean;
  isSaved?: boolean;
}
export interface CustomerMinRealOwnerDataClient {
  curp: string;
  foreignerWithoutCurp: boolean;
  rfc: string;
  firstName: string;
  middleName: string;
  dateOfBirth: string;
  firstLastName: string;
  secondLastName: string;
  nationality: string;
  countryOfBirth: string;
  countryTaxCodeAbroad?: string;
  typeIden: string;
}